+++ src/components/vice/admin/table/index.js
@@ -1,7 +1,6 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 
 import {
-    build as buildID,
     withI18N,
     EnhancedTableHead,
     getMessage as msg,
@@ -20,166 +19,16 @@ import {
     IconButton,
     useTheme,
     useMediaQuery,
-    Button,
-    Popper,
 } from "@material-ui/core";
-
-import { Skeleton } from "@material-ui/lab";
 
 import { KeyboardArrowUp, KeyboardArrowDown } from "@material-ui/icons";
 
-import { useQuery } from "react-query";
-
-import {
-    asyncData,
-    ASYNC_DATA_QUERY_KEY,
-} from "../../../../serviceFacades/vice/admin";
+import ActionButtons from "./actionButtons";
 
 import messages from "./messages";
 import ids from "./ids";
+import { id } from "./functions";
 import useStyles from "./styles";
-
-// Constructs an ID for an element.
-const id = (...names) => buildID(ids.BASE, ...names);
-
-const ActionButtonsSkeleton = () => {
-    return (
-        <Skeleton variant="rect" animation="wave" height={75} width="100%" />
-    );
-};
-
-const ActionButton = ({ baseID, name, handler, onClick, popperMsgKey }) => {
-    const classes = useStyles();
-    return (
-        <Button
-            id={id(baseID, "button", name)}
-            variant="contained"
-            color="primary"
-            onClick={(event) => {
-                event.preventDefault();
-                event.stopPropagation();
-                return onClick(event, handler, popperMsgKey);
-            }}
-            className={classes.actionButton}
-        >
-            {msg(name)}
-        </Button>
-    );
-};
-
-const ActionButtons = ({
-    row,
-    handleExtendTimeLimit = (_) => {},
-    handleDownloadInputs = (_) => {},
-    handleUploadOutputs = (_) => {},
-    handleExit = (_) => {},
-    handleSaveAndExit = (_) => {},
-}) => {
-    const classes = useStyles();
-
-    const [anchorEl, setAnchorEl] = useState(null);
-    const [open, setOpen] = useState(false);
-    const [popperMessage, setPopperMessage] = useState("");
-
-    const { status, data, error } = useQuery(
-        [ASYNC_DATA_QUERY_KEY, row.externalID],
-        asyncData
-    );
-
-    const isLoading = status === "loading";
-    const hasErrored = status === "error";
-
-    if (hasErrored) {
-        console.log(error);
-    }
-
-    const onClick = (event, dataFn, msgKey) => {
-        let tlErr;
-        let tlData;
-
-        try {
-            tlData = dataFn(data.analysisID, row.externalID);
-        } catch (err) {
-            tlErr = err;
-        }
-
-        setAnchorEl(event.currentTarget);
-        setPopperMessage(tlErr ? tlErr.message : msg(msgKey));
-        setOpen(true);
-
-        return tlData;
-    };
-
-    useEffect(() => {
-        const timerID = setInterval(() => {
-            if (open) {
-                setOpen(false);
-            }
-        }, 3000);
-        return () => clearInterval(timerID);
-    });
-
-    return (
-        <div className={classes.actions}>
-            {isLoading ? (
-                <ActionButtonsSkeleton />
-            ) : (
-                <>
-                    <ActionButton
-                        baseID={row.externalID}
-                        name="extendTimeLimit"
-                        handler={handleExtendTimeLimit}
-                        popperMsgKey="timeLimitExtended"
-                        onClick={onClick}
-                    />
-
-                    <ActionButton
-                        baseID={row.externalID}
-                        name="downloadInputs"
-                        handler={handleDownloadInputs}
-                        popperMsgKey="downloadInputsCommandSent"
-                        onClick={onClick}
-                    />
-
-                    <ActionButton
-                        baseID={row.externalID}
-                        name="uploadOutputs"
-                        handler={handleUploadOutputs}
-                        popperMsgKey="uploadOutputsCommandSent"
-                        onClick={onClick}
-                    />
-
-                    <ActionButton
-                        baseID={row.externalID}
-                        name="exit"
-                        handler={handleExit}
-                        popperMsgKey="exitCommandSent"
-                        onClick={onClick}
-                    />
-
-                    <ActionButton
-                        baseID={row.externalID}
-                        name="saveAndExit"
-                        handler={handleExit}
-                        popperMsgKey="saveAndExitCommandSent"
-                        onClick={onClick}
-                    />
-
-                    <Popper
-                        id={id(row.externalID, "popper")}
-                        open={open}
-                        anchorEl={anchorEl}
-                        placement="top"
-                    >
-                        <div className={classes.paperPopper}>
-                            {popperMessage}
-                        </div>
-                    </Popper>
-                </>
-            )}
-        </div>
-    );
-};
 
 const ExtendedDataCard = ({
     columns,
